Fix edit & delete test
